requirements.txt
fastapi
uvicorn
feedparser
numpy
faiss-cpu
sentence-transformers
requests